User Routes added

+++ src/services/cars.services.ts
@@ -2,19 +2,35 @@ import "reflect-metadata";
 import { injectable } from "tsyringe";
 import { TCarSchema, TCreateCar, TUpdateCarBody } from "../schemas/cars.schemas";
 import { prisma } from "../database/prisma";
+import { AppError } from "../erros/AppError";
 
 
 @injectable()
 export class CarServices {
 
-    async createCar(body: TCreateCar): Promise<TCarSchema> {
+    async createCar(body: TCreateCar, userId: string): Promise<TCarSchema> {
 
-        const newCar = await prisma.car.create({ data: body })
+        const newCar = await prisma.car.create({
+            data: {
+                ...body,
+                userId
+            }
+        })
 
         return newCar
     }
 
-    async getManyCars(): Promise<TCarSchema[]> {
+    async getManyCars(userId?: string): Promise<TCarSchema[]> {
+
+        if (userId) {
+            const userCars = await prisma.car.findMany({
+                where: {
+                    userId: userId
+                }
+            })
+
+            return userCars
+        }
 
         const data = await prisma.car.findMany()
 
@@ -30,17 +46,44 @@ export class CarServices {
         return data
     }
 
-    async updateCar(carId: string, body: TUpdateCarBody): Promise<TCarSchema> {
+    async updateCar(carId: string, body: TUpdateCarBody, userId: string): Promise<TCarSchema> {
 
-        const data = await prisma.car.update({
+        const existingCar = await prisma.car.findUnique({
+            where: { id: carId }
+        });
+
+        if (!existingCar) {
+            throw new AppError(404, "Car not found");
+        }
+
+        if (existingCar.userId !== userId) {
+            throw new AppError(404, "You are not the owner of this car");
+        }
+
+        const updatedCar = await prisma.car.update({
             where: { id: carId },
-            data: body
-        })
+            data: {
+                ...body,
+                userId: existingCar.userId
+            }
+        });
 
-        return data
+        return updatedCar;
     }
 
-    async deleteCars(carId: string): Promise<void> {
+    async deleteCars(carId: string, userId: string): Promise<void> {
+
+        const existingCar = await prisma.car.findUnique({
+            where: { id: carId }
+        });
+
+        if (!existingCar) {
+            throw new AppError(404, "Car not found");
+        }
+
+        if (existingCar.userId !== userId) {
+            throw new AppError(404, "You are not the owner of this car");
+        }
 
         await prisma.car.delete({
             where: { id: carId }
